pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hgattn"]

[project]
name = "strange-loop-quest"
version = "0.1.0"
description = "Experiments to summon the Strange Loop"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fire>=0.7.1",
    "hydra-core>=1.3.2",
    "music21>=9.9.1",
    "numpy>=2.4.2",
    "torch",
]
[project.optional-dependencies]
cu124 = [
  "torch",
  "torchvision"
]
cu130 = [
  "torch",
  "torchvision"
]


[tool.uv]
package = true
conflicts = [
  [
    { extra = "cu124" },
    { extra = "cu130" },
  ],
]

[tool.uv.sources]
torchvision = [
  { index = "pytorch-cu130", extra = "cu130" },
  { index = "pytorch-cu124", extra = "cu124" },
]
torch = [
  { index = "pytorch-cu130", extra = "cu130" },
  { index = "pytorch-cu124", extra = "cu124" },
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

